PythonBot/requirements.txt
paho-mqtt==1.4.0
google-api-core==1.13.0
google-auth==1.6.3
googleapis-common-protos==1.6.0
dialogflow==0.6.0
gTTS==2.0.3
playsound==1.2.2